Split api-handlers into separate routes
