+++ .gitignore
@@ -4,6 +4,7 @@
 /.next
 /out
 next-env.d.ts
+*.tsbuildinfo
 
 # Dependencies
 /node_modules
