+++ pages/api/checkAuth.ts
@@ -1,9 +1,10 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { getDb, getConfig } from '../../server-src/services';
-import { checkAuthHandler } from '../../server-src/api-handlers';
+import { getConfig } from '../../server-src/services';
+import { apiWrapper } from '../../server-src/api-handlers';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const db = await getDb();
   const config = await getConfig();
-  return checkAuthHandler(db, config, req as any, res as any);
+  await apiWrapper(req as any, res as any, config, 'GET checkAuth', () => {
+    res.status(200).send('success');
+  });
 }
